+++ spec/bundle_outdated_formatter/formatter/tsv_formatter_spec.rb
@@ -1,0 +1,56 @@
+RSpec.describe BundleOutdatedFormatter::TSVFormatter do
+  let(:pretty) { false }
+  let(:formatter) { described_class.new(pretty: pretty) }
+
+  let(:outdated_gems) do
+    [
+      {
+        'gem'       => 'faker',
+        'newest'    => '1.6.6',
+        'installed' => '1.6.5',
+        'requested' => '~> 1.4',
+        'groups'    => 'development, test'
+      },
+      {
+        'gem'       => 'hashie',
+        'newest'    => '3.4.6',
+        'installed' => '1.2.0',
+        'requested' => '= 1.2.0',
+        'groups'    => 'default'
+      },
+      {
+        'gem'       => 'headless',
+        'newest'    => '2.3.1',
+        'installed' => '2.2.3',
+        'requested' => '',
+        'groups'    => ''
+      }
+    ]
+  end
+
+  let(:text_tsv) do
+    <<-EOS.chomp
+"gem"	"newest"	"installed"	"requested"	"groups"
+"faker"	"1.6.6"	"1.6.5"	"~> 1.4"	"development, test"
+"hashie"	"3.4.6"	"1.2.0"	"= 1.2.0"	"default"
+"headless"	"2.3.1"	"2.2.3"	""	""
+    EOS
+  end
+
+  describe '#convert' do
+    before do
+      formatter.instance_variable_set(:@outdated_gems, outdated_gems)
+    end
+
+    subject { formatter.convert }
+
+    context 'when @pretty is false' do
+      it { is_expected.to eq text_tsv }
+    end
+
+    context 'when @pretty is true' do
+      let(:pretty) { true }
+      it { is_expected.to eq text_tsv }
+    end
+  end
+end